clip_check_recursive is not related to picking
